refactor: restructure project layout, apply UI style updates, and initialize RAG integration setup
