+++ dashboard/app_drift/pages/embedding_visualization.py
@@ -24,8 +24,8 @@ def render():
         return
 
     embedding_data = st.session_state['embedding_data']
-    dataset_name = st.session_state.get('dataset_name', 'Dataset')
-    st.title(f"Embedding Visualization of {dataset_name}")
+    dataset_name = st.session_state.get('dataset_name')
+    st.write(f"Embedding Visualization of {dataset_name}")
 
     # 데이터셋 분리 함수
     def extract_embeddings(key):
