--- dashboard/docs/DriftDetection-Tools.md
@@ -1,189 +1,0 @@
-# **Open-Source Drift Detection Tools** (수정 중)
-
----
-
-머신러닝 모델 성능 저하 원인인 데이터 드리프트 탐지를 위해 세가지 오픈소스인 `evidentlyai`,  `nannyml`,  `alibi-detect` 를 알아보고 비교
-
-## **1. [EvidentlyAI](https://github.com/evidentlyai/evidently)**
-
-> 데이터 드리프트 감지와 ML 품질 검토, 모니터링을 포함한 다양한 기능을 갖춘 도구
->
-
-### 1.1 특징
-
-Machine Learning, Large Language Models 의 평가와 모니터링을 위한 오픈소스 Python 라이브러리로 AI system과 Data pipeline을 실험 단계부터 프로덕션까지 평가, 테스트, 모니터링할 수 있도록 지원
-
-- **장점**
-    - **다양한 메트릭 제공**
-        
-        데이터와 모델 성능을 평가할 수 있는 다양한 프리셋 및 메트릭을 제공하여 설정이 간단하고, 결과를 바로 시각화할 수 있어 분석이 편리
-        
-    - **리포트와 대시보드 기능**
-        
-        직관적인 리포트 생성과 대시보드를 통해 시각화된 결과를 확인
-        
-    - **손쉬운 통합**
-        
-        다양한 데이터 드리프트 감지 기능을 제공하며, 분류와 회귀 작업의 성능 모니터링에 유리
-        
-- **단점**
-    - **임베딩 비교 기능 부족**
-        
-        임베딩 기반의 드리프트 감지 기능은 제한적이며, 사용자가 별도로 벡터화와 임베딩 분석을 수행
-        
-    - **텍스트 데이터에 대한 제한적 지원**
-        
-        텍스트 데이터를 벡터화하거나, 한국어와 같은 다국어 텍스트 데이터를 다루기 위해서는 사전 처리 과정이 필요
-        
-
-### 1.2 드리프트 감지
-
-Statistical Test, Distance Metric을 통해 모델 예측, 수치형 및 범주형 특성, 텍스트 데이터, 임베딩의 분포 변화를 비교
-
-기준 데이터인 reference 데이터셋과 현재 데이터인 current 데이터셋의 열을 비교해 각 열의 유형과 관측 수, 고유 값의 수를 고려해 적합한 드리프트 테스트를 선택
-
-→ reference는 train data, current는 test data를 사용하여 비교
-
-각 데이터의 특징에 따라 데이터 드리프트 탐지하는 방법이 달라짐
-
-- 수치형 데이터
-    - 고유 값이 많은 수치형 열은 Kolmogorov-Smirnov Test 또는 Wasserstein Distance
-- 범주형 데이터
-    - 고유 값이 적은 범주형 열은 Chi-Squared Test 또는
-- 텍스트 데이터
-    - 도메인 분류기를 사용해 기준 데이터와 현재 데이터의 분포 차이를 분류 모델로 판단
-    - 도메인 분류기 domain classifier : 두 데이터셋 간 분포 차이를 식별화하기 위해 훈련된 이진 분류 모델로 두 데이터셋 소속 여측
-- 임베딩
-    - ROC AUC 기준으로 드리프트 여부를 감지하고 필요시 임베딩 간 거리를 추가로 활용 가능
-
-### 1.3 약점
-
-- **토크나이저**
-    
-    한국어에 맞춘 토크나이징이나 형태소 분석 지원이 없어서 전처리 기능 미비
-    
-- **임베딩 거리 메트릭 자체를 계산하는 기능 없음**
-    
-    사용자가 직접 Word2Vec, TF-IDF 또는 사전 훈련된 언어 모델을 이용해 직접 텍스트 데이터를 벡터로 변환 후, 해당 오픈소스에 전달
-    
-- **언어모델 제한**
-    
-    드리프트 감지 시, 한국어 텍스트 의미를 잘 반영하기 위해 KoBERT, KoELECTRA 같은 한국어 모델 기반의 벡터화가  필요
-    
-- **특정 문화적 맥락이나 특수 어휘 감지 미흡**
-    
-    한국어 특유의 어휘나 표현 방식, 문장 길이나 특수 문자를 다루는 부분에서도 문화적 차이를 반영하지 못함
-
-## **2. [NannyML](https://github.com/NannyML/nannyml)**
-
-> 모델 성능 평가와 데이터 드리프트 감지, 그리고 예측 정확성에 미치는 영향을 평가하는 데에 강점
-> 
-
-### 2.1 특징
-시계열 데이터에서 발생하는 데이터 드리프트를 감지하고 모니터링하는 데에 유용한 기능을 갖춤
-
-시계열 데이터의 ‘청크(chunk)’ 단위로 데이터를 나누어 드리프트를 감지하는 기능을 제공 
-
-→ 변화의 정확한 시기를 파악하고 예측 정확도에 미치는 영향을 평가하는 데 탁월
-
-- 장점
-    - **모델 성능 예측**
-        
-        실제 레이블 없이도 모델 성능을 추정할 수 있어, 특히 모델이 배포된 환경에서 레이블이 항상 제공되지 않을 때 유용
-        
-    - **Feature Level의 세밀한 드리프트 추적**
-        
-        피처별 드리프트 모니터링이 가능하며, 모델 성능에 미치는 영향까지 분석
-        
-    - **강력한 텍스트 데이터 지원**
-        
-        텍스트 데이터 특성을 포함한 드리프트 감지 기능이 있어, 텍스트 데이터의 분포 변화를 감지하고 관리하기 좋음
-        
-- **단점**
-    - **설정 및 사용 난이도**
-        
-        다소 설정이 복잡할 수 있고, 모델 성능 추정 기능을 정확히 활용하려면 사전 지식이 필요
-        
-    - **텍스트 임베딩 직접 지원 부족**
-        
-        텍스트 임베딩과의 통합은 지원하지 않으며, 임베딩 벡터를 사전 생성하여 입력 필요
-
-### 2.2 드리프트 감지
-### 
-
-네 가지 감지 알고리즘이 있으며, 임계값 설정 방식을 통해 알림을 조정
-
-사용자가 임계값을 직접 설정하거나, 표준편차를 기반으로 임계값을 자동으로 설정
-
-- Univariate Drift Detection 단변량 드리프트 탐지
-    
-    각 변수별로 개별 드리프트 확인하고, 분석 데이터 기간동안 생성된 chunk를 reference data 기간과 비교해 drift 계산
-    
-    - 연속형, 범주형 데이터에 사용할 방법 지정, 날짜는 타임스탬프 컬럼을 지정해 시계열 분석 사용
-    - 적합을 통해 모델 학습한 뒤 드리프트 계산
-    - 각 결과는 메서드를 사용해 시각화 진행
-
-- Multivariate Drift Detection 다변량 드리프트 탐지
-    
-    데이터 구조의 미묘한 변화를 탐지해 단변량 방법으로는 감지할 수 없는 변화를 포착
-    
-    하나의 요약값으로 결과를 제공해 잘못된 경고의 위험 감소하나 단변량에 비해 설명 가능성이 낮음
-
-### 2.3 약점 
-주로 시계열 데이터와 수치형 데이터에 적합한 데이터 드리프트 감지 툴
-
-## **3. [Alibi-Detect](https://github.com/SeldonIO/alibi-detect)**
-
-> 이상치, 데이터 드리프트, 적대적 공격 탐지를 위해 설계 되었으며 TensorFlow, PyTorch 등 다양한 백엔드와 호환되는 도구
-> 
-
-### 3.1 특징
-- 장점
-    - **텍스트 및 임베딩 데이터에 강점**
-        
-        텍스트 데이터와 임베딩 벡터의 드리프트를 감지하기 위한 다양한 통계적 테스트(KL Divergence, JS Divergence 등)를 기본적으로 지원
-        
-    - **유연한 확장성**
-        
-        다양한 임베딩 방법과 분포 비교 기법을 사용해 텍스트 데이터를 포함한 복잡한 데이터 유형에서도 효과적으로 드리프트를 분석
-        
-- 단점
-    - **사용자 설정 복잡성**
-        
-        설정이 상대적으로 복잡할 수 있으며, 특정 메트릭과 메소드 설정에 대한 사전 지식이 필요
-        
-    - **시각화 및 리포트 제한**
-        
-        Evidently처럼 포괄적인 시각화 리포트는 지원하지 않기 때문에, 시각화 및 분석에는 추가 작업이 필요
-### 3.2 드리프트 감지
-**다양한 Drift Detection 방법 제공**
-
-- 가설검정
-    
-    특정 통계치 기준으로 두 분포가 같은지 여부를 평가하며 유의미한 차이가 있는 경우 드리프트 발생
-    
-- Univariate Tests
-    
-    특성별로 독립적인 단변량 검정 수행
-    
-- Multivariate Tests
-    
-    다변량 검정을 통해 모든 특성을 고려해 드리프트 감지 ex. MMD, LSDD
-
-### 3.4 약점 
-분석 간결 : 자동화된 리포트는 따로 없음
-
-## 요약
-
-** 각 공식 문서 참고
-
-|  | 탐지 범위 | 강점 | 특징 |
-| --- | --- | --- | --- |
-| **EvidentlyAI** | Feature Drift | 일반적인 데이터 드리프트 감지 | 초기 분석 단계(시각화 가능) |
-| **NannyML** | Feature Drift, Concept Drift | 변화의 정확한 시기를 파악 | 모델 성능의 사후 평가(production 환경에 유용) |
-| **Alibi**-**Detect** | Covariate Drift, Concept Drift, Label Drift | 고차원 데이터(image, text) | 미세 변화 탐지 |
-
-### 드리프트 탐지 이후,
-
-- 데이터 드리프트만을 모니터링하는 것은 불완전
-- 성능 추정 지표를 함께 추적해 모델 성능에 대한 완전한 이해가 가능
