+++ dashboard/milvus_db/inspect-collections.py
