+++ dashboard/milvus_db/rm-collections.py
